Extensive file, namespace, class renaming.
