+++ joint_limits_interface/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(safety_limits_interface)
+project(joint_limits_interface)
 
 if(USE_ROSBUILD)
 
@@ -9,12 +9,12 @@ if(USE_ROSBUILD)
   set(EXECUTABLE_OUTPUT_PATH ${PROJECT_SOURCE_DIR}/bin)
   set(LIBRARY_OUTPUT_PATH ${PROJECT_SOURCE_DIR}/lib)
 
-  rosbuild_add_gtest(soft_joint_limits_interface_test  test/soft_joint_limits_interface_test.cpp)
-  rosbuild_add_gtest(joint_limits_urdf_test            test/joint_limits_urdf_test.cpp)
+  rosbuild_add_gtest(joint_limits_interface_test test/joint_limits_interface_test.cpp)
+  rosbuild_add_gtest(joint_limits_urdf_test      test/joint_limits_urdf_test.cpp)
 
   # TODO: why is it explicitly needed???, without it the linker fails.
-  target_link_libraries(soft_joint_limits_interface_test  pthread)
-  target_link_libraries(joint_limits_urdf_test            pthread)
+  target_link_libraries(joint_limits_interface_test pthread)
+  target_link_libraries(joint_limits_urdf_test      pthread)
 
 else()
 
@@ -26,8 +26,8 @@ else()
     INCLUDE_DIRS include
     )
 
-  catkin_add_gtest(soft_joint_limits_interface_test  test/soft_joint_limits_interface_test.cpp)
-  catkin_add_gtest(joint_limits_urdf_test            test/joint_limits_urdf_test.cpp)
+  catkin_add_gtest(joint_limits_interface_test test/joint_limits_interface_test.cpp)
+  catkin_add_gtest(joint_limits_urdf_test      test/joint_limits_urdf_test.cpp)
 
   # Install
   install(DIRECTORY include/${PROJECT_NAME}/
