+++ joint_limits_interface/package.xml
@@ -1,7 +1,7 @@
 <package>
-  <name>safety_limits_interface</name>
+  <name>joint_limits_interface</name>
   <version>1.0.0</version>
-  <description>Hardware Interaface base class.</description>
+  <description>Interface for enforcing joint limits.</description>
   <maintainer email="adolfo.rodriguez@pal-robotics.com">Adolfo Rodriguez Tsouroukdissian</maintainer>
   <license>BSD</license>
 
@@ -13,9 +13,11 @@
 
   <build_depend>roscpp</build_depend>
   <build_depend>hardware_interface</build_depend>
+  <build_depend>urdf_interface</build_depend>
 
   <run_depend>roscpp</run_depend>
   <run_depend>hardware_interface</run_depend>
+  <build_depend>urdf_interface</build_depend>
 
   <export>
     <cpp cflags="-I${prefix}/include"/>
